split out chatbot recs
